rename pendinghandler folder to pending
